backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
paho-mqtt==2.0.0
aiofiles==23.2.1
python-multipart==0.0.9
